+++ pumpkin/src/block/blocks/redstone/repeater.rs
@@ -8,7 +8,7 @@ use pumpkin_data::{
 use pumpkin_macros::pumpkin_block;
 use pumpkin_util::math::position::BlockPos;
 use pumpkin_world::world::BlockFlags;
-use pumpkin_world::{BlockStateId, chunk::TickPriority};
+use pumpkin_world::{BlockStateId, tick::TickPriority};
 
 use crate::{
     block::{
@@ -237,9 +237,9 @@ impl RedstoneGateBlock<RepeaterProperties> for RepeaterBlock {
         }
     }
 
-    fn get_update_delay_internal(&self, state_id: BlockStateId, block: &Block) -> u16 {
+    fn get_update_delay_internal(&self, state_id: BlockStateId, block: &Block) -> u8 {
         let props = RepeaterProperties::from_state_id(state_id, block);
-        (props.delay.to_index() + 1) * 2
+        (props.delay.to_index() as u8 + 1) * 2
     }
 }
